refactor: move conf.py into configuration dir and split it intofee and network
