+++ crypto/identity/wif.py
@@ -4,7 +4,7 @@ from base58 import b58encode_check
 
 from binary.unsigned_integer.writer import write_bit8
 
-from crypto.conf import get_network
+from crypto.configuration.network import get_network
 
 
 def wif_from_passphrase(passphrase, network_wif=None):
